% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.qmatrix}
\alias{as.qmatrix}
\title{Converts into Q matrix}
\usage{
as.qmatrix(Q)
}
\arguments{
\item{Q}{an object of class \code{matrix} containing a matrix of ancestry coefficients.}
}
\value{
An object of class \code{tess3Q}.
}
\description{
This function tests Q matrix objects, and converts \code{matrix} objects into valid Q matrices.
}
\examples{
library(tess3r)
## an example with 3 individuals and 2 clusters
Qmatrix <- matrix(c(0.4,0.6,0.3,0.7, 0.2, 0.8), byrow = TRUE, nrow = 3)
Qmatrix <- as.qmatrix(Qmatrix)
barplot(Qmatrix, space = 0, xlab = "individuals",
        ylab = "Ancestry proportions", main = "Ancestry matrix")
}
\seealso{
\code{\link{plot.tess3Q}} \code{\link{barplot.tess3Q}}
}
\author{
Kevin Caye, Olivier François
}
