% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{barplot.tess3Q}
\alias{barplot.tess3Q}
\title{structure barplot for Q-matrix objects.}
\usage{
\method{barplot}{tess3Q}(Q, col.palette = NULL, palette.step = 9, ...)
}
\arguments{
\item{plot.type}{}
}
\description{
structure barplot for Q-matrix objects.
}

