% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.R
\name{CalculateEmpiricalGenSemivariogram}
\alias{CalculateEmpiricalGenSemivariogram}
\title{Calculate empirical semi variance from genotype matrix.}
\usage{
CalculateEmpiricalGenSemivariogram(X, ploidy, coord, breaks = "FD",
  na.rm = TRUE)
}
\arguments{
\item{X}{Genotype matrix.}

\item{ploidy}{The number of chromosome.}

\item{coord}{Coordinate matrix.}

\item{breaks}{Same parameter that in hist R base function.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\value{
Semi-variance
}
\description{
Calculate empirical semi variance from genotype matrix.
}
\examples{
library(tess3r)

data("data.for.test", package = "tess3r")
em.vario <- CalculateEmpiricalGenSemivariogram(X = data.for.test$X,
                                ploidy = 1,
                                coord = data.for.test$coord,
                                breaks = "FD",
                                na.rm = TRUE)
ggplot2::ggplot(em.vario, ggplot2::aes(x = h, y = semi.variance, size = size)) +
               ggplot2::geom_point()
}
