% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckCoord}
\alias{CheckCoord}
\title{Raise error if coord is not well formated.}
\usage{
CheckCoord(coord)
}
\arguments{
\item{coord}{Coordinate matrix.}
}
\value{

}
\description{
Raise error if coord is not well formated.
}

