% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckWCoord}
\alias{CheckWCoord}
\title{Raise error if W and coord are not consistent.}
\usage{
CheckWCoord(W, coord)
}
\arguments{
\item{W}{Graph weight matrix.}

\item{coord}{Coordinate matrix.}
}
\description{
Raise error if W and coord are not consistent.
}
