% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckX}
\alias{CheckX}
\title{Raise error if X is not well formated.}
\usage{
CheckX(X, ploidy)
}
\arguments{
\item{X}{Genotype matrix.}

\item{ploidy}{The number of chromosome.}
}
\description{
Raise error if X is not well formated.
}
