% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckXBin}
\alias{CheckXBin}
\title{Raise error if XBin is not well formated.}
\usage{
CheckXBin(XBin, ploidy)
}
\arguments{
\item{XBin}{Binary genotype matrix.}

\item{ploidy}{The number of chromosome.}
}
\value{

}
\description{
Raise error if XBin is not well formated.
}

