% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckXBinCoord}
\alias{CheckXBinCoord}
\title{Raise error if XBin and coord are not consistent.}
\usage{
CheckXBinCoord(XBin, coord)
}
\arguments{
\item{XBin}{Binary genotype matrix.}

\item{coord}{Coordinate matrix.}
}
\description{
Raise error if XBin and coord are not consistent.
}
