% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckXBinW}
\alias{CheckXBinW}
\title{Raise error if XBin and W are not consistent.}
\usage{
CheckXBinW(XBin, W)
}
\arguments{
\item{XBin}{Binary genotype matrix.}

\item{W}{Graph weight matrix.}
}
\description{
Raise error if XBin and W are not consistent.
}
