% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckXCoord}
\alias{CheckXCoord}
\title{Raise error if X and coord are not consistent.}
\usage{
CheckXCoord(X, coord)
}
\arguments{
\item{X}{Genotype matrix.}

\item{coord}{Coordinate matrix.}
}
\value{

}
\description{
Raise error if X and coord are not consistent.
}

