% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{CheckXW}
\alias{CheckXW}
\title{Raise error if X and W are not consistent.}
\usage{
CheckXW(X, W)
}
\arguments{
\item{X}{Genotype matrix.}

\item{W}{Graph weight matrix.}
}
\description{
Raise error if X and W are not consistent.
}

