% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ComputeChi2AndPvalue}
\alias{ComputeChi2AndPvalue}
\title{Convert Fst into chi2 and compute p value}
\usage{
ComputeChi2AndPvalue(Fst, K, n)
}
\arguments{
\item{Fst}{Matrix of Fst.}

\item{K}{Number of ancestral populations.}

\item{n}{Number of individuals.}
}
\description{
Convert Fst into chi2 and compute p value
}

