% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ComputeEigenValuesWithRSpectra}
\alias{ComputeEigenValuesWithRSpectra}
\title{Compute eigen values with RSpectra.}
\usage{
ComputeEigenValuesWithRSpectra(Lapl, k)
}
\arguments{
\item{Lapl}{Graph laplacian matrix.}

\item{k}{Number of eigen values.}
}
\description{
Compute eigen values with RSpectra.
}
