% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeExponetialWeight}
\alias{ComputeExponetialWeight}
\title{Compute graph matrix with an exponiential function.}
\usage{
ComputeExponetialWeight(coord, sigma)
}
\arguments{
\item{coord}{Coordinate matrix.}

\item{sigma}{sigma numeric parameter.}
}
\description{
Compute graph matrix with an exponiential function.
}

