% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeFst}
\alias{ComputeFst}
\title{Compute Fst statistic}
\usage{
ComputeFst(Q, G, D)
}
\arguments{
\item{Q}{Ancestry coefficients matrix.}

\item{G}{Ancestral frequencies matrix.}

\item{D}{Number of chromosome + 1.}
}
\description{
Compute Fst statistic
}

