% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ComputeGraphBasedOnVariogram}
\alias{ComputeGraphBasedOnVariogram}
\title{Title}
\usage{
ComputeGraphBasedOnVariogram(X, coord, plot = TRUE, nugget = NULL, ...)
}
\arguments{
\item{plot}{}
}
\description{
Title
}

