% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ComputeGraphLaplacian}
\alias{ComputeGraphLaplacian}
\title{Compute laplacian of a graph.}
\usage{
ComputeGraphLaplacian(W)
}
\arguments{
\item{W}{Graph weight matrix.}
}
\description{
Compute laplacian of a graph.
}

