% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeHeatKernelWeight}
\alias{ComputeHeatKernelWeight}
\title{Compute graph matrix with the heat kernel function.}
\usage{
ComputeHeatKernelWeight(coord, sigma)
}
\arguments{
\item{coord}{Coordinate matrix.}

\item{sigma}{sigma numeric parameter.}
}
\description{
Compute graph matrix with the heat kernel function.
}
