% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeHeatKernelWeightSparse}
\alias{ComputeHeatKernelWeightSparse}
\title{Compute graph sparse matrix with the heat kernel function.}
\usage{
ComputeHeatKernelWeightSparse(coord, sigma)
}
\arguments{
\item{coord}{Coordinate matrix.}

\item{sigma}{sigma numeric parameter.}
}
\description{
Compute graph sparse matrix with the heat kernel function.
}

