% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeMCPASolution}
\alias{ComputeMCPASolution}
\title{solve min || X - Q G^T|| + lambda * tr(Q^T Lapl Q)}
\usage{
ComputeMCPASolution(X, K, Lapl, lambdaPrim, D, maxIteration, tolerance, Q, G)
}
\description{
solve min || X - Q G^T|| + lambda * tr(Q^T Lapl Q)
}

