% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeMCPASolutionNoCopyX}
\alias{ComputeMCPASolutionNoCopyX}
\title{solve min || X - Q G^T|| + lambda * tr(Q^T Lapl Q)}
\usage{
ComputeMCPASolutionNoCopyX(X, K, Lapl, lambdaPrim, D, maxIteration, tolerance,
  Q, G)
}
\description{
solve min || X - Q G^T|| + lambda * tr(Q^T Lapl Q)
}

