% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ComputeMeanDist}
\alias{ComputeMeanDist}
\title{Compute the average distance between points.}
\usage{
ComputeMeanDist(coord)
}
\arguments{
\item{coord}{Coordinate matrix.}
}
\description{
Compute the average distance between points.
}

