% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeRmse}
\alias{ComputeRmse}
\title{Compute root squared mean error}
\usage{
ComputeRmse(Q1, Q2)
}
\arguments{
\item{Q1}{Numeric.}

\item{Q2}{Numeric.}
}
\description{
Compute root squared mean error
}
