% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{ComputeRmseWithBestPermutationGreedy}
\alias{ComputeRmseWithBestPermutationGreedy}
\title{Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse. Warning : Greedy algorithm !}
\usage{
ComputeRmseWithBestPermutationGreedy(Q1, Q2)
}
\arguments{
\item{Q2}{Numeric.}
}
\description{
Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse. Warning : Greedy algorithm !
}

