% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{ComputeSpatialPenalty}
\alias{ComputeSpatialPenalty}
\title{Compute the value spatial penalty used in TESS3 objectif function.}
\usage{
ComputeSpatialPenalty(Q, Lapl)
}
\arguments{
\item{Lapl}{}
}
\description{
Compute the value spatial penalty used in TESS3 objectif function.
}

