% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{CreatePalette}
\alias{CreatePalette}
\title{Create a list of palettes}
\usage{
CreatePalette(color.vector = c("tomato", "chartreuse", "gold", "blue",
  "violet", "wheat", "olivedrab"), palette.length = 9)
}
\arguments{
\item{color.vector}{a vector of R colors.}

\item{palette.length}{an integer number of colors in each palette.}
}
\value{
an object of class \code{list} containing a list of color palettes.
}
\description{
This function creates a list of color palettes for the plot and barplot
functions
}
\examples{
library(tess3r)

## Load A. thaliana example
data(data.at)
obj <- tess3(data.at$X, coord = data.at$coord, K = 5, ploidy = 1,
             openMP.core.num = 4)
Qmatrix <- qmatrix(obj,K=5)
my.colors <- c("tomato", "yellow", "blue", "wheat","olivedrab")
my.palette <- CreatePalette(my.colors, 9)
plot(obj$Q, data.at$coord, method = "mapping.max", col.palette = my.palette,
     cex = .4, xlab = "Longitude", ylab= "Latitude",
     main = "Ancestry coefficients")
}
\author{
Kevin Caye, Olivier François
}
\seealso{
\code{\link{plot.tess3Q}} \code{\link{barplot.tess3Q}}
}

