% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{FieldsKrigModel}
\alias{FieldsKrigModel}
\title{Return an interpolation function. Return a wrapper function of the function Krig of
the package fields. See \code{\link[fields]{Krig}}.}
\usage{
FieldsKrigModel(theta = 10, ...)
}
\arguments{
\item{theta}{Numeric.}

\item{...}{Parameters of \code{\link[fields]{Krig}}.}
}
\description{
Return an interpolation function. Return a wrapper function of the function Krig of
the package fields. See \code{\link[fields]{Krig}}.
}
