% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{FieldsTpsModel}
\alias{FieldsTpsModel}
\title{Return an interpolation function. Return a wrapper function of the function Tps of
the package fields. See \code{\link[fields]{Tps}}.}
\usage{
FieldsTpsModel(...)
}
\arguments{
\item{...}{Parameters of \code{\link[fields]{Tps}}.}
}
\description{
Return an interpolation function. Return a wrapper function of the function Tps of
the package fields. See \code{\link[fields]{Tps}}.
}
