% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{Gettess3res}
\alias{Gettess3res}
\title{Get tess3 run result}
\usage{
Gettess3res(tess3, K, rep = "best")
}
\arguments{
\item{tess3}{A tess3 object.}

\item{K}{Number of ancestral populations.}

\item{rep}{Repetition index.}
}
\value{
An object of class \code{\link{tess3Main}}.
}
\description{
Get the \code{\link{tess3Main}} object returned by the TESS3 algorithm.
}
\seealso{
\code{\link{tess3}}
}

