% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{ggtess3Q}
\alias{ggtess3Q}
\title{Displays geographic maps for ancestry coefficients}
\usage{
ggtess3Q(Q, coord, resolution = c(300, 300), window = NULL,
  background = TRUE, map.polygon = NULL, raster.filename = NULL,
  interpolation.model = FieldsKrigModel(10), col.palette = CreatePalette())
}
\arguments{
\item{Q}{An object of class \code{tess3Q}. The ancestry coefficient matrix.}

\item{coord}{The numeric matrix of size \eqn{n \times 2} where \eqn{n} is the
number of individuals.}

\item{resolution}{An integer vector of the resolution of the grid used to
computed the interpolating surface}

\item{window}{The window size, such that \code{window = c(xmin, xmax, ymin, ymax)}
contains the window mina nd max coordinates.}

\item{background}{If TRUE the raster file is used as a stencil to render only
raster pixel on earth.}

\item{map.polygon}{The \code{sp::SpatialPolygonsDataFrame} object used to crop the interpolating surfaces.
If \code{NULL}, the function \code{\link[rworldmap]{getMap}} is used.}

\item{raster.filename}{The raster file name used to compute the background stencil.
This is an alternative method to crop the interpolating surfaces. The default method uses \code{map.polygon}.}

\item{interpolation.model}{The interpolation model used to compute compute
the interpolating surface. You can use functions \code{\link{FieldsTpsModel}},
\code{\link{FieldsKrigModel}}.}

\item{col.palette}{A list of color palettes. You can use the function \code{\link{CreatePalette}}.}
}
\value{
None
}
\description{
Displays geographic maps for ancestry coefficients.
}
\examples{
library(tess3r)

# Retrieve a dataset
data(data.at)

# Run of TESS3
obj <- tess3(data.at$X, coord = data.at$coord, K = 5,
                 ploidy = 1, method = "projected.ls", openMP.core.num = 4)

# Get the ancestry matrix
Q.matrix <- qmatrix(obj, K = 5)

# Plot the spatial interpolation of the ancestry matrix
ggtess3Q(Q.matrix, data.at$coord)
}
\author{
Kevin Caye, Flora Jay, Olivier François
}

