% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionscan.R
\name{GtoFreq}
\alias{GtoFreq}
\title{Compute allele frquencies matrix from the genotype frequencie matrix.}
\usage{
GtoFreq(G, ploidy)
}
\arguments{
\item{G}{Genotype frequencie matrix.}

\item{ploidy}{The number of chromosome.}
}
\value{
TODOC
}
\description{
Compute allele frquencies matrix from the genotype frequencie matrix.
}

