% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{idw}
\alias{idw}
\title{Return an interpolation function. Return a wrapper function of the function idw of
the package gstat. See \code{\link[gstat]{idw}}}
\usage{
idw(idp = 1)
}
\arguments{
\item{idp}{numeric; specify the inverse distance weighting power.}
}
\description{
Return an interpolation function. Return a wrapper function of the function idw of
the package gstat. See \code{\link[gstat]{idw}}
}

