% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ImputeRound}
\alias{ImputeRound}
\title{Impute missing value with the Q ancestry coeficient matrix and G ancestral
frequencies estimates. Missing values are computed as a round of estimated genotype
frequencies (P = Q G^T).}
\usage{
ImputeRound(tess3.res, masked.X)
}
\arguments{
\item{tess3.res}{tess3Main object with Q and G estimates.}

\item{masked.X}{Genotype matrix with the missing values (NA values).}
}
\description{
Impute missing value with the Q ancestry coeficient matrix and G ancestral
frequencies estimates. Missing values are computed as a round of estimated genotype
frequencies (P = Q G^T).
}

