% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{is.tess3}
\alias{is.tess3}
\title{Test if x is a tess3project object}
\usage{
is.tess3(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if x is an tess3 object.
}
\description{
Test if x is a tess3project object
}

