% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{kriging}
\alias{kriging}
\title{Return an interpolation function. Return a wrapper function of the function Krig of
the package fields. See \code{\link[fields]{Krig}}.}
\usage{
kriging(theta = 10)
}
\arguments{
\item{theta}{Numeric.}
}
\description{
Return an interpolation function. Return a wrapper function of the function Krig of
the package fields. See \code{\link[fields]{Krig}}.
}

