% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{PlotInterpotationAll}
\alias{PlotInterpotationAll}
\title{Plot all the interpolation surfaces on several graphs.}
\usage{
PlotInterpotationAll(coord, list.grid.z, grid.x, grid.y, col.palette, map, ...)
}
\arguments{
\item{coord}{Coordinate matrix.}

\item{list.grid.z}{List of interpolation surface matrices.}

\item{grid.x}{TODOC}

\item{grid.y}{TODOC}

\item{col.palette}{List of color palette.}

\item{map}{If true map function of maps package is call to plot polygon from
map database.}

\item{...}{TODOC}
}
\description{
Plot all the interpolation surfaces on several graphs.
}

