% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{pvalue}
\alias{pvalue}
\title{Get pvalue vector.}
\usage{
pvalue(tess3, K, rep = "best")
}
\arguments{
\item{tess3}{A tess3 object.}

\item{K}{Number of ancestral populations.}

\item{rep}{Repetition index.}
}
\value{
The \code{pvalue} ancestry coefficient matrix.
}
\description{
Get the \code{pvalue} vector returned by the TESS3 algorithm.
}
\seealso{
\code{\link{tess3}}
}

