% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{RasterAncestryCoefRgb}
\alias{RasterAncestryCoefRgb}
\title{TODO}
\usage{
RasterAncestryCoefRgb(Q, coord,
  raster.filename = system.file("extdata/raster", "earth.tif", package =
  "TESS3enchoSen"), interpolation.function = idw(), resolution = list(ncol =
  300, nrow = 300))
}
\description{
TODO
}

