% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{rmse.tess3Main}
\alias{rmse.tess3Main}
\title{Title}
\usage{
rmse.tess3Main(tess3.obj, X, ploidy, mask = NULL)
}
\arguments{
\item{tess3.obj}{TODOCC}

\item{X}{TODOC}

\item{ploidy}{TODOC}

\item{mask}{TODOC}
}
\description{
Title
}
