% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleDistFromCenterDirichletQ}
\alias{SampleDistFromCenterDirichletQ}
\title{sample pop center and compute dist.from.center as dirichlet parameter}
\usage{
SampleDistFromCenterDirichletQ()
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

