% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleFuncQ}
\alias{SampleFuncQ}
\title{sample Q = f(coord)}
\usage{
SampleFuncQ(coord, f = function(X) c(1/(1 + exp(-0.5 * X[1])), 1 - 1/(1 +
  exp(-0.5 * X[1]))))
}
\arguments{
\item{K}{}
}
\description{
sample Q = f(coord)
}

