% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleGaussianProcessQ}
\alias{SampleGaussianProcessQ}
\title{sample Q such Q = logit(a)
a ~ N(0,W) W_ij = cov.function(X_i,X_j)}
\usage{
SampleGaussianProcessQ(cov.function)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

