% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleGenoFromGenerativeModel}
\alias{SampleGenoFromGenerativeModel}
\title{Sample X such that P(X_i_dl + j) = Sum(Q_ik G_kj).
Coord ...}
\usage{
SampleGenoFromGenerativeModel(n.by.pop, L, d, K,
  G.sampler = SampleDirichletG(), Q.sampler = SampleUnifQ(),
  coord.sampler = SampleNormalClusterCoord(),
  locus.position.sampler = SampleUnifLocusPosition())
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

