% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleGenoOFWithMs}
\alias{SampleGenoOFWithMs}
\title{Title}
\usage{
SampleGenoOFWithMs(n, nsites.neutral, nsites.selected, crossover.proba,
  m.neutral, m.selected, mutation.rate.per.site, N0 = 10^6, k = 0.5,
  min.maf = 0.05, plot.debug = FALSE, tess3.ms = getOption("tess3.ms"))
}
\arguments{
\item{n}{number of indivudual to sample}

\item{nsites.neutral}{number of site between which recombination occur for neutral loci}

\item{nsites.selected}{number of site between which recombination occur for selected loci}

\item{crossover.proba}{corss-over probability between adjacent site per generation}

\item{m.neutral}{migration rate for neutral loci}

\item{m.selected}{migration rate for selected loci}

\item{mutation.rate.per.site}{mutation rate per site}

\item{N0}{population size}

\item{min.maf}{the locus with a maf less than this parameter are removed}

\item{plot.debug}{if TRUE plot at different stage of the simulation}
}
\description{
Title
}
\examples{

tess3.ms <- "~/BiocompSoftware/msdir/ms"
n <- 200
K <- 2
ploidy <- 1
data.list <- SampleGenoOFWithMs(n = n,
                               nsites.neutral = 100000,
                               nsites.selected = 1000,
                               crossover.proba = 0.25 * 10 ^ -8,
                               m.neutral = 0.25 * 10 ^ -6,
                               m.selected = 0.25 * 10 ^ -7,
                               mutation.rate.per.site = 0.25 * 10 ^ -8,
                               N0 = 10 ^ 6,
                               k = 0.5,
                               min.maf = 0.05,
                               plot.debug = TRUE)


}

