% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleTESS2.3Q}
\alias{SampleTESS2.3Q}
\title{sample Q as describe in "Spatial Inference of Admixture Proportions and Segondary Contact Zones" E. Durand et al.}
\usage{
SampleTESS2.3Q(coord, K, W, sigma, rho, f = function(X) c(1, X[1], X[2]),
  beta = matrix(0, 3, K))
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

