% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleUnifDirichletG}
\alias{SampleUnifDirichletG}
\title{sample G such as G_dl+._k ~ Dirichlet(1/(ploidy + 1))}
\usage{
SampleUnifDirichletG(L, ploidy, K)
}
\arguments{
\item{L}{TODOC}

\item{ploidy}{TODOC}

\item{K}{TODOC}
}
\value{
TODOC
}
\description{
TODOC
}
