% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleUnifQ}
\alias{SampleUnifQ}
\title{sample Q such as alpha ~ unif(0,1) and Q ~ Dirichlet(alpha)}
\usage{
SampleUnifQ(n, K)
}
\arguments{
\item{K}{}
}
\description{
sample Q such as alpha ~ unif(0,1) and Q ~ Dirichlet(alpha)
}

