% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{summary.tess3Main}
\alias{summary.tess3Main}
\title{Summary of tess3 object.}
\usage{
\method{summary}{tess3Main}(object, ...)
}
\arguments{
\item{object}{TODOC}

\item{...}{TODOC}
}
\value{
TODOCC
}
\description{
Summary of tess3 object.
}

