% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{tess3project}
\alias{tess3project}
\title{Run tess3 for multiple value of K and with repetition.}
\usage{
tess3project(X, coord, K, ploidy, lambda, rep = 1, W = NULL,
  method = "MCPA", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, keep = "all")
}
\arguments{
\item{mask}{}
}
\description{
Run tess3 for multiple value of K and with repetition.
}

