% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erics-funcs.R
\name{tess3Q_map_rasters}
\alias{tess3Q_map_rasters}
\title{Return rasters underlying the geographic maps of ancestry coefficients}
\usage{
tess3Q_map_rasters(x, coord, method = "map.max", resolution = c(300, 300),
  window = NULL, background = TRUE, map.polygon = NULL,
  raster.filename = NULL, interpolation.model = FieldsKrigModel(10),
  col.palette = CreatePalette(), ...)
}
\arguments{
\item{x}{an object of class \code{tess3Q} containing an ancestry coefficient matrix computed from \code{\link{tess3}}.}

\item{coord}{a numeric matrix of dimension \eqn{n} times 2 where \eqn{n} is the
number of individuals. The matrix must contain (Longitude, Latitude) coordinates for all individuals.}

\item{method}{THIS DOESN'T HAVE ANY EFFECT HERE...THE rasters that come back are the same.  a character string \code{"map.max"} or \code{"map.all"}. If \code{"map.all"}, interpolating surfaces are displayed
for the ancestry coefficients of all populations. If \code{"map.max"} the union of interpolating surfaces is displayed.}

\item{resolution}{an integer vector \code{resolution = c(rx,ry)} for the resolution of the grid used to
computed the interpolating surface. \code{rx} and \code{ry} are resolution numbers for the x-axis and y-axis respectively.}

\item{window}{an integer vector for the plotting window, such that \code{window = c(xmin, xmax, ymin, ymax)}
contains the window's min and max coordinates.}

\item{background}{if \code{TRUE} a raster file is used as a stencil to render only raster pixel on earth.}

\item{map.polygon}{an object of class \code{sp::SpatialPolygonsDataFrame} used to crop the interpolating surface.
If \code{NULL}, the function \code{\link[rworldmap]{getMap}} is used.}

\item{raster.filename}{a raster file name used to compute the background stencil.
This is an alternative method to crop the interpolating surfaces. The default method uses \code{map.polygon}.}

\item{interpolation.model}{an interpolation model used to compute the interpolating surface. Interpolation models can use the
functions \code{\link{FieldsTpsModel}} or \code{\link{FieldsKrigModel}}.}

\item{col.palette}{a list of color palettes. Color palettes can be defined by using the function \code{\link{CreatePalette}}.}

\item{...}{\code{\link{plot.default}} other parameters.}
}
\value{
None
}
\description{
This function returns the rasterStack that underlies the plots that you would get
if you were to send the same arguments to the function \code{\link{plot.tess3Q}},
'cuz it might be nice to get at the underlying plot data.
}
\examples{
library(tess3r)

# Retrieve a dataset
data(data.at)

# Run of TESS3
obj <- tess3(data.at$X, coord = data.at$coord, K = 5,
                 ploidy = 1, method = "projected.ls", openMP.core.num = 4)

# Get the ancestry matrix
Q.matrix <- qmatrix(obj, K = 5)

# Plot the spatial interpolation of the ancestry matrix
boing <- tess3Q_map_rasters(Q.matrix, data.at$coord, method = "map.max",
     resolution = c(400,400),
     interpolation.model = FieldsKrigModel(10), cex = .4,
     xlab = "Longitude", ylab= "Latitude", main = "Ancestry coefficients")
}
\author{
Kevin Caye, Flora Jay, Olivier François (repackaged by Eric C. Anderson)
}

