% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\docType{package}
\name{tess3rExperiment}
\alias{tess3rExperiment}
\alias{tess3rExperiment-package}
\title{tess3r : estimation of spatial population structure}
\description{
This R package implements the TESS3 method and tools useful to plot program outputs.
}

