% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{TestRequiredPkg}
\alias{TestRequiredPkg}
\title{Test if a package is installed.}
\usage{
TestRequiredPkg(pkg)
}
\arguments{
\item{pkg}{Package name to test.}
}
\description{
Test if a package is installed.
}

