% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{universalkriging}
\alias{universalkriging}
\title{Return an interpolation function. Return a wrapper function of the function autoKrige of
the package automap.}
\usage{
universalkriging(formula = as.formula(Z ~ X + Y))
}
\arguments{
\item{formula}{Formula that defines the dependent variable as a linear model of independent variables.}
}
\description{
Return an interpolation function. Return a wrapper function of the function autoKrige of
the package automap.
}

