% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{X2XBin}
\alias{X2XBin}
\title{Binary coding of the genotype data matrix.}
\usage{
X2XBin(X, ploidy, XBin)
}
\arguments{
\item{X}{Genotype matrix.}

\item{ploidy}{Numbet of chromosomes.}

\item{XBin}{Binary genotype matrix to fill.}
}
\description{
Binary coding of the genotype data matrix.
}
