% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XBin2X}
\alias{XBin2X}
\title{Convert the binary genotype matrix into genotype matrix.}
\usage{
XBin2X(XBin, ploidy)
}
\arguments{
\item{XBin}{Binary genotype matrix.}

\item{ploidy}{Numbet of chromosomes.}
}
\description{
Convert the binary genotype matrix into genotype matrix.
}
