% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{barplot.tess3Q}
\alias{barplot.tess3Q}
\title{Barplot representation of a Q-matrix}
\usage{
\method{barplot}{tess3Q}(Q, sort.by.Q = TRUE, col.palette = NULL,
  palette.length = 9, ...)
}
\arguments{
\item{Q}{an object of class \code{tess3Q} (Q matrix) containing a matrix of ancestry coefficients computed from \code{tess3} or converted from other program formats.}

\item{sort.by.Q}{a Boolean value indicating whether individuals should be sorted by their ancestry level or not.}

\item{col.palette}{is a list of color palettes. If \code{NULL}, a default list with 8 color palettes is used.}

\item{palette.length}{an integer value for the number of colors in each element of the palette list.}
}
\value{
Generates a graphical output.
}
\description{
This function displays a barplot representation of the
ancestry coefficient matrix. It includes a sort-by-Q option.
}
\examples{
data(data.at)
obj <- tess3(data.at$X, coord = data.at$coord, K = 5, ploidy = 1, openMP.core.num = 4)
Qmatrix <- obj$Q
barplot(Qmatrix, border = NA, space = 0, xlab = "individuals", ylab = "Ancestry proportions", main = "Ancestry matrix")
}
\author{
Kevin Caye, Olivier François
}
\seealso{
\code{\link{plot.tess3Q}} \code{\link{as.qmatrix}} \code{\link{CreatePalette}}
}

