% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{barplot.tess3Q}
\alias{barplot.tess3Q}
\title{Barplot representation of a Q-matrix}
\usage{
\method{barplot}{tess3Q}(height, sort.by.Q = TRUE, col.palette = NULL,
  palette.length = 9, lab = FALSE, ...)
}
\arguments{
\item{height}{an object of class \code{tess3Q} (Q matrix) containing a matrix of ancestry coefficients computed from \code{tess3} or converted from other program formats.}

\item{sort.by.Q}{a Boolean value indicating whether individuals should be sorted by their ancestry level or not.}

\item{col.palette}{is a list of color palettes. If \code{NULL}, a default list with 8 color palettes is used.}

\item{palette.length}{an integer value for the number of colors in each element of the palette list.}

\item{lab}{a list of individual labels.}

\item{...}{other parameters of the function \code{\link{barplot.default}}.}
}
\value{
A permutation of individual labels used in the sort.by.Q option (order). Displays the
Q matrix.
}
\description{
This function displays a barplot representation of the
ancestry coefficient matrix. It includes a sort-by-Q option.
}
\examples{
library(tess3r)

# Retrieve a dataset
data(data.at)

# Run of TESS3
obj <- tess3(data.at$X, coord = data.at$coord, K = 5,
                 ploidy = 1, method = "projected.ls", openMP.core.num = 4)

# Get the ancestry matrix
Q.matrix <- qmatrix(obj, K = 5)

# Display a barplot for the Q matrix
barplot(Q.matrix, border = NA, space = 0, xlab = "Individuals",
        ylab = "Ancestry proportions", main = "Ancestry matrix") -> bp
axis(1, at = 1:nrow(Q.matrix), labels = bp$order, las = 3, cex.axis = .4)
}
\author{
Kevin Caye, Olivier François
}
\seealso{
\code{\link{plot.tess3Q}} \code{\link{as.qmatrix}} \code{\link{CreatePalette}}
}

