% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{barplot.tess3Q}
\alias{barplot.tess3Q}
\title{barplot for Q-matrix objects.}
\usage{
\method{barplot}{tess3Q}(Q, sort.by.Q = TRUE, col.palette = NULL,
  palette.step = 9, ...)
}
\arguments{
\item{plot.type}{}
}
\description{
barplot for Q-matrix objects.
}

