% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.R
\name{CalculateEmpiricalSemivariogram}
\alias{CalculateEmpiricalSemivariogram}
\title{Title}
\usage{
CalculateEmpiricalSemivariogram(X, ploidy, coord, breaks = "FD",
  na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
Title
}
\examples{
library(tess3r)

data("data.for.test", package = "tess3r")
em.vario <- empirical.variogram(X = data.for.test$X,
                                ploidy = 1,
                                coord = data.for.test$coord,
                                breaks = "FD",
                                na.rm = TRUE)
ggplot2::ggplot(em.vario, ggplot2::aes(x = h, y = semi.variance, size = size)) +
               ggplot2::geom_point()
}

