% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{ComputeAveragedCrossEntropy}
\alias{ComputeAveragedCrossEntropy}
\title{Compute the averaged cross entropy mean(-P * log(Q))}
\usage{
ComputeAveragedCrossEntropy(P, Q)
}
\description{
Compute the averaged cross entropy mean(-P * log(Q))
}

