% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeAveragedCrossEntropy}
\alias{ComputeAveragedCrossEntropy}
\title{Compute average cross entropy}
\usage{
ComputeAveragedCrossEntropy(P, Q)
}
\arguments{
\item{P}{Numeric.}

\item{Q}{Numeric.}
}
\description{
Compute average cross entropy
}

