% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ComputeTscoreAndPvalue}
\alias{ComputeTscoreAndPvalue}
\title{Convert Fst into t-score and compute p value}
\usage{
ComputeTscoreAndPvalue(Fst, K, n)
}
\description{
Convert Fst into t-score and compute p value
}

