% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionscan.R
\name{GtoFreq}
\alias{GtoFreq}
\title{Compute allele frequency matrix from genotype frequency matrix.}
\usage{
GtoFreq(G, ploidy)
}
\arguments{
\item{G}{genotype frequency matrix.}

\item{ploidy}{the number of chromosomes.}
}
\value{
a matrix of allele frequencies for each locus.
}
\description{
Compute allele frequency matrix from genotype frequency matrix.
}

