% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{plot.tess3}
\alias{plot.tess3}
\title{Plot RMSE(X, Q * t(G)) for all K number of ancestral population with error bars.}
\usage{
\method{plot}{tess3}(object, crossvalid = FALSE, crossentropy = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot RMSE(X, Q * t(G)) for all K number of ancestral population with error bars.
}

