% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{plot.tess3}
\alias{plot.tess3}
\title{Plot cross-validation errors for all values of number of ancestral populations}
\usage{
\method{plot}{tess3}(x, crossvalid = FALSE, crossentropy = FALSE, ...)
}
\arguments{
\item{x}{a tess3 object.}

\item{crossvalid}{if TRUE, errors are evaluated on masked data.}

\item{crossentropy}{If TRUE, the cross-entropy error is used. If FALSE,
the root mean square error is used.}

\item{...}{other graphic parameters.}
}
\description{
Plot cross-validation errors for all values of number of ancestral populations
}
