% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotG.R
\name{plot.tess3pvalue}
\alias{plot.tess3pvalue}
\title{Plot the manhattan plot of -log(p.value).}
\usage{
\method{plot}{tess3pvalue}(x, ...)
}
\arguments{
\item{x}{tess3pvalue object.}

\item{...}{TODOC}
}
\value{
TODOC
}
\description{
Plot the manhattan plot of -log(p.value).
}

