% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotG.R
\name{plot.tess3pvalue}
\alias{plot.tess3pvalue}
\title{Manhattan plot for tess3 significance values.}
\usage{
\method{plot}{tess3pvalue}(x, ...)
}
\arguments{
\item{x}{tess3pvalue object.}

\item{...}{\code{\link{plot.default}} other parameters.}
}
\value{
none
}
\description{
Manhattan plot for tess3 significance values.
}
\seealso{
See \code{\link{tess3Main}} more explainations and examples.
}
