% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{plot.tess3Q}
\alias{plot.tess3Q}
\title{Plot individual ancestry coefficients on a map.}
\usage{
\method{plot}{tess3Q}(Q, coord, plot.type = "piechart", resolution = c(300,
  300), window = NULL, background = TRUE, raster.filename = NULL,
  interpolation.function = idw(), col = NULL, col.palette = NULL,
  map = TRUE, palette.step = 9, ...)
}
\arguments{
\item{plot.type}{}
}
\description{
Plot individual ancestry coefficients on a map.
}

