% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlots.R
\name{PlotPiechartAncestryCoef}
\alias{PlotPiechartAncestryCoef}
\title{Title}
\usage{
PlotPiechartAncestryCoef(Q, coord, window, background, col,
  radius = sqrt((window[2] - window[1])^2 + (window[4] - window[3])^2) * 0.01,
  ...)
}
\arguments{
\item{window}{}
}
\description{
Title
}

