% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{pvalue}
\alias{pvalue}
\title{Get p-values from a genome scan for selection}
\usage{
pvalue(tess3, K, rep = "best")
}
\arguments{
\item{tess3}{an object of class \code{tess3}.}

\item{K}{a number of ancestral populations.}

\item{rep}{an integer correponding to a run number (default is 'best' run).}
}
\value{
Corrected pvalues for all loci to be used in a scan for adaptive alleles.
}
\description{
Get the \code{pvalue} vector returned by the tess3 algorithm.
}
\seealso{
\code{\link{tess3}}
}
