% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{qmatrix}
\alias{qmatrix}
\title{Get Q matrix estimate.}
\usage{
qmatrix(tess3, K, rep = "best")
}
\arguments{
\item{tess3}{A tess3 object.}

\item{K}{Number of ancestral populations.}

\item{rep}{Repetition index.}
}
\description{
Get Q matrix estimate.
}

