% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{run.ms}
\alias{run.ms}
\title{Helper function which call ms.}
\usage{
run.ms(ms.file, nsam, nreps, theta, rho, nsites, M)
}
\arguments{
\item{ms.file}{TODOC}

\item{nsam}{TODOC}

\item{nreps}{TODOC}

\item{theta}{TODOC}

\item{rho}{TODOC}

\item{nsites}{TODOC}

\item{M}{TODOC}
}
\value{
TODOC
}
\description{
Helper function which call ms.
}
