% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleDistFromCenterDirichletQ}
\alias{SampleDistFromCenterDirichletQ}
\title{For population the center mu_k is computed. Then D_i_k = ||coord_i - mu_k|| is computed. finally alpha = f(D) and Q ~ Dirichlet(alpha).}
\usage{
SampleDistFromCenterDirichletQ(coord, n.by.pop, K, f = function(D)
  exp(-D/0.2))
}
\arguments{
\item{coord}{Coordinate matrix.}

\item{n.by.pop}{Number of individual by population.}

\item{K}{Number of column of Q.}

\item{f}{Distribution function.}
}
\value{
TODO
}
\description{
TODO
}

