% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleGenoFromGenerativeModelTESS3}
\alias{SampleGenoFromGenerativeModelTESS3}
\title{Sample X such that P(X_i_dl + j) = Sum(Q_i_k G_k_dl + j).}
\usage{
SampleGenoFromGenerativeModelTESS3(Q, G, coord, ploidy)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

