% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleNormalClusterCoord}
\alias{SampleNormalClusterCoord}
\title{sample coord such as a mixture of K cluster distributed with gaussian law}
\usage{
SampleNormalClusterCoord()
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

