% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SampleUnifQ}
\alias{SampleUnifQ}
\title{Sample ancestral coefficient matrix such as alpha ~ unif(0,1) and Q ~ Dirichlet(alpha).}
\usage{
SampleUnifQ(n, K)
}
\arguments{
\item{n}{Number of row of Q.}

\item{K}{Number of column of Q.}
}
\description{
Sample ancestral coefficient matrix such as alpha ~ unif(0,1) and Q ~ Dirichlet(alpha).
}

