% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tess2tess3}
\alias{tess2tess3}
\title{Import input files from STRUCTURE and TESS 2.3}
\usage{
tess2tess3(dataframe = NULL, TESS = TRUE, diploid = TRUE, FORMAT = 1,
  extra.row = 0, extra.column = 0)
}
\arguments{
\item{dataframe}{a data frame read from a STRUCTURE or a TESS file.}

\item{TESS}{a boolean value set to \code{TRUE} if the TESS 2.3 format is used. If \code{TRUE}, the
geographic coordinates (Longitude, Latitude) must be binded left to the genetic marker matrix.}

\item{diploid}{a boolean value set to \code{TRUE} for diploids and \code{FALSE} for haploids.}

\item{FORMAT}{an integer value equal to 1 for markers encoded using one row of data for each individual, and
\code{FORMAT = 2} for markers encoded using two rows of data for each individual.}

\item{extra.row}{an integer value indicating the number of extra rows in the header of the input file (marker ids).}

\item{extra.column}{an integer value indicating the number of extra columns in the input file. Extra columns come before the geographic coordinates
in TESS 2.3 input files, and include individual ids, pop ids, phenotypes, etc. Geographic coordinates must be considered as extra columns if the flag \code{TESS}
is set to \code{TESS = FALSE}. Missing data are encoded as "-9" or any negative values}
}
\value{
An object of class \code{list} containing a genotype matrix and individual geographic coordinates.

X a matrix of genotypes

coord a matrix of geographic coordinates
}
\description{
This function converts data sets imported from the STRUCTURE and the TESS 2.3 input file formats to the tess3 formats.
}
\examples{
##
}
\author{
Kevin Caye, Flora Jay, Olivier François
}
\seealso{
\code{\link{tess3}}
}

