% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{tess3}
\alias{tess3}
\title{Main function.}
\usage{
tess3(X, XBin = NULL, coord, K, ploidy, lambda = 1, W = NULL,
  method = "MCPA", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, no.copy = FALSE)
}
\arguments{
\item{no.copy}{if TRUE data will not be duplicate in memory but computation can take more time}
}
\description{
Main function.
}

