% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{tess3}
\alias{tess3}
\title{Run tess3 for multiple value of K and with repetition.}
\usage{
tess3(X, XBin = NULL, coord, K, ploidy, lambda = 1, rep = 1, W = NULL,
  method = "MCPA", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, copy = TRUE,
  algo.copy = TRUE, keep = "best", verbose = FALSE)
}
\arguments{
\item{no.copy}{}
}
\description{
Run tess3 for multiple value of K and with repetition.
}

