% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{tess3}
\alias{tess3}
\title{Estimation of ancestry proportions and genome scans for selection}
\usage{
tess3(X, XProba = NULL, coord, K, ploidy, lambda = 1, rep = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, algo.copy = TRUE,
  keep = "best", verbose = FALSE)
}
\arguments{
\item{X}{a matrix of individual genotypes. This matrix must
have \eqn{n} rows and  \eqn{L} columns where \eqn{n} is the number of individuals and
\eqn{L} is the number of loci. The entries of this matrix are integers between 0 and
ploidy, that correspond to the number of derived/reference alleles observed at each locus.
If \code{NULL}, the matrix of genotype likelihood, \code{XProba}, is used.}

\item{XProba}{A matrix which contains individual genotype likelihoods (probabilities) for each
locus. This matrix must contain \eqn{n} rows and \eqn{(ploidy + 1)L} columns where
\eqn{n} is the number of individuals, and \eqn{L} is the number of loci. The entries of
this matrix are numeric values ranging between 0 and 1, and corresponding to genotype
probabilities for each locus. If \code{NULL}, deterministic values are computed from the
genotypic matrix \code{X}. See the references for more details.}

\item{coord}{a matrix of size \eqn{n \times 2} where \eqn{n} is the
number of individuals encoding the longitude and latitude of each individual (numeric system).}

\item{K}{an integer or a vector of integers corresponding to
the number of ancestral populations.}

\item{ploidy}{an integer corresponding to ploidy of the studied
organism. Haploids have ploidy = 1, diploids have ploidy = 2, etc.}

\item{lambda}{a numeric value for the spatial regularization parameter.
The default value lambda = 1 attributes equal weights to the loss function
and to the penalty function.}

\item{rep}{integer. The number of time the algorithm will be repeated for each value of
\code{K}.}

\item{W}{a matrix which corresponds to the graph weightings.
If NULL, W is computed as
\code{W[i,j] = exp( - (coord[i] - coord[j])^2 / sigma^2)},
where \code{coord[i]} represents
the geographic coordinates for individual i, and where
\code{sigma} is equal to 5 percent of the average geographic distance between individuals.}

\item{method}{a character string \code{"projected.ls"} or \code{"qp"}. If \code{"projected.ls"},
an alternating projected least squares algorithm is used. If \code{"qp"},
an alternating quadratic programing algorithm is used. See references for more
details}

\item{max.iteration}{the maximum number of iterations of the optimization algorithm.}

\item{tolerance}{a numeric value corresponding to the
stopping criteria of the optimization algorithm.}

\item{openMP.core.num}{integer representing the number of cores used by the optimization
algorithm. It requires that the openMP library is installed in your OS (default for macOS is no).}

\item{Q.init}{a matrix for initial values of ancestry coefficients for the algorithm. The
default value is a random matrix.}

\item{mask}{If not \code{NULL}, this numeric value is the proportion of masked data
when computing the cross-validation criterion.}

\item{algo.copy}{boolean. If TRUE data is copied to speed up the algorithm.}

\item{keep}{If \code{"best"}, only the result with the lowest \code{rmse} score will be kept
for each value of \code{K}. If \code{"all"}, all results will be kept
and returned for each value of \code{K}. The second option uses more space in memory.}

\item{verbose}{If \code{TRUE} more information is printed.}
}
\value{
An object of class tess3 which corresponds to a list of length \code{length(K)}.
Each element of this list has the following attributes
\describe{
   \item{K}{the number of ancestral populations}
   \item{tess3.run}{if \code{keep = "best"}, the \code{\link{tess3Main}} result
   with the lowest value of the \code{rmse} (loss) function. If \code{keep = "all"},
   a list of \code{\link{tess3Main}} results for each repetition}
   \item{rmse}{root mean squared error between the genotypic matrix \code{XProba} and the
   fitted matrix for each program repetition}
   \item{crossentropy}{cross-entropy between the genotypic matrix \code{XProba} and the
   fitted matrix for each program repetition}
   \item{crossvalid.rmse}{root square mean error between the masked values of genotypic matrix
    \code{XProba[masked]} and their fitted values for each repetition. If mask is FALSE, then \code{NULL}.}
   \item{crossvalid.crossentropy}{cross-entropy between the masked values of genotypic matrix
    \code{XProba[masked]} and their fitted values for each repetition. If mask is FALSE, then \code{NULL}.}
}
Methods available for this class:
\itemize{
  \item \code{\link{plot.tess3}}
  \item \code{\link{summary.tess3}}
  \item \code{\link{is.tess3}}
  \item \code{\link{Gettess3res}}
  \item \code{\link{qmatrix}}
  \item \code{\link{pvalue}}
}
}
\description{
\code{tess3} is the main function of the \code{tess3r} package. It runs
a graph-based nonnegative matrix factorization algorithm that includes
geographic data in the estimation of spatial population structure.
The function requires individual genotypes, geographic coordinates, and
it can be run for multiple values of the number of ancestral populations.
In addition, the function uses the estimates of ancestry coefficients to
compute an Fst statistic for each locus, and to return test significance
values for a null hypothesis of selective neutrality. See the references
for more details.
}
\examples{
library(tess3r)

# Arabidopsis thaliana data set
data(data.at)
genotype <- data.at$X
coordinates <- data.at$coord

# Running the tess3 function
tess3.obj <- tess3(X = genotype, coord = coordinates, K = 1:4,
                   method = "projected.ls",
                   ploidy = 1)

# Plot error
plot(tess3.obj, pch = 19, col = "blue",
     xlab = "Number of ancestral populations",
     ylab = "Cross-validation score")

# Retrieve the Q-matrix for K = 3 clusters
q.matrix <- qmatrix(tess3.obj, K = 3)

## STRUCTURE-like barplot for the Q-matrix
barplot(q.matrix, border = NA, space = 0,
       xlab = "Individuals", ylab = "Ancestry proportions",
       main = "Ancestry matrix") -> bp
axis(1, at = 1:nrow(q.matrix), labels = bp$order, las = 3, cex.axis = .4)

## Spatial interpolation of ancestry coefficient
my.colors <- c("tomato", "orange", "lightblue")
my.palette <- CreatePalette(my.colors, 9)
plot(q.matrix, coordinates, method = "map.max", interpol = kriging(10),
     main = "Ancestry coefficients",
     xlab = "Longitude", ylab = "Latitude",
     resolution = c(500,500), cex = .4,
     col.palette = my.palette)

## Genome scan p-values for K = 3
p.values <- pvalue(tess3.obj, K = 3)
hist(p.values, col = "lightblue")

## Manhatan plot
plot(p.values, main = "Manhattan plot",
    xlab = "Locus id",
    ylab = "-log10(P-values)",
    cex = .3, col = "grey")

}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1111/1755-0998.12471/full}
Caye, Kevin et al. (2016) Fast Inference of Individual Admixture Coefficients Using Geographic Data. bioRxiv
doi:10.1101/080291. \url{http://biorxiv.org/content/early/2016/10/12/080291}
}
\seealso{
\code{\link{tess3Main}}, \code{\link{plot.tess3Q}},
\code{\link{barplot.tess3Q}}
}

