% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{tess3}
\alias{tess3}
\title{Run tess3 for multiple value of K and with repetition.}
\usage{
tess3(X, XProba = NULL, coord, K, ploidy, lambda = 1, rep = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, algo.copy = TRUE,
  keep = "best", verbose = FALSE)
}
\arguments{
\item{X}{TODOC}

\item{coord}{TODOC}

\item{K}{TODOC}

\item{ploidy}{TODOC}

\item{lambda}{TODOC}

\item{rep}{TODOC}

\item{W}{TODOC}

\item{method}{TODOC}

\item{max.iteration}{TODOC}

\item{tolerance}{TODOC}

\item{openMP.core.num}{TODOC}

\item{Q.init}{TODOC}

\item{mask}{TODOC}

\item{algo.copy}{If TRUE data will be copy to speed the algorithm.}

\item{keep}{TODOC}

\item{verbose}{TODOC}

\item{XBin}{TODOC}

\item{copy}{TODOC}
}
\value{
An object of class tess3.
}
\description{
Run tess3 for multiple value of K and with repetition.
}

