% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{tess3}
\alias{tess3}
\title{Main function.}
\usage{
tess3(genotype, geographic.coordinate, K, ploidy, lambda, W = NULL,
  method = "MCPA", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL)
}
\arguments{
\item{tolerance}{}
}
\description{
Main function.
}

