% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{tess3Main}
\alias{tess3Main}
\title{Estimates spatial population structure}
\usage{
tess3Main(X, XProba = NULL, coord, K, ploidy, lambda = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, copy = TRUE,
  algo.copy = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{A numeric matrix which corresponds to the genotype matrix. This matrix
must be of size \eqn{n \times L} where \eqn{n} is the number of individual and
\eqn{L} is the number of loci. Values of this matrix are integer corresponding
to the number of variant alleles observed at a locus. If \code{NULL}, \code{XProba}
is used.}

\item{XProba}{A numeric matrix which correspond to the probability for the genotype.
This matrix must be of size \eqn{n \times (ploidy + 1)L} where
\eqn{n} is the number of individual, \eqn{L} is the number of loci. Values of
this matrix are numeric between 0 and 1 corresponding
to the genome probability. It is the matrix used in graph based non negative
factorization matrix. If \code{NULL}, it is computed from the genotype matrix \code{X}.
See reference for more details.}

\item{coord}{The numeric matrix of size \eqn{n \times 2} where \eqn{n} is the
number of individuals.}

\item{K}{An integer which corresponds to
the number of ancestral populations.}

\item{ploidy}{An integer which corresponds to
the number of set of chromosomes.}

\item{lambda}{A numeric which corresponds to the
spatial regularization parameter.}

\item{W}{A numeric matrix which corresponds to the graph weiht matrix.
If NULL, W it is computed as
\code{W[i,j] = exp( - (coord[i] - coord[j])^2 / sigma^2)}.
Where \code{coord[i]} is
the geographic coordinate for the individual i and
\code{sigma} equals 5 percent of the average geographic distance between individuals.}

\item{method}{\code{"projected.ls"} or \code{"qp"}. If \code{"projected.ls"},
an aleternated projected least squares algorithm is used. If \code{"qp"},
an alternated quadratic programing algorithm is used. See references for more
details}

\item{max.iteration}{The max number of
iteration of the optimization algorithm.}

\item{tolerance}{A numeric which corresponds to the tolerance of the
stopping criteria of the optimization algorithm.}

\item{openMP.core.num}{If openMP is available on your computer, it is the
number of core used by the algorithm.}

\item{Q.init}{A numeric matrix which corresponds to the initial value of
\code{Q} for the algorithm.}

\item{mask}{If not \code{NULL} this the proportion of the data matrix which
is masked to compute the cross validation criteria.}

\item{copy}{if TRUE data will be copy 1 time.}

\item{algo.copy}{if TRUE data will be copy 1 time to speed the algorithm.}

\item{verbose}{If \code{TRUE} more information are printed.}
}
\value{
An object of class tess3Main which is a list with components:
\describe{
   \item{L}{The number of loci.}
   \item{n}{The number of individuals.}
   \item{ploidy}{The number of set of chromosomes.}
   \item{K}{The number of ancestral population.}
   \item{G}{The ancestral genotype frequency matrix.}
   \item{Q}{The ancestry coefficient matrix.}
   \item{Fst}{The Fst statistic computed for each locus.}
   \item{Fscore}{The Fscore computed from Fst.}
   \item{pvalue}{The pvalues computed from Fscore.}
   \item{log.pvalue}{The \eqn{log(pvalue)}.}
   \item{rmse}{The root square mean error between \code{XProba} and
   \code{tcrossprod(Q, G)}.}
   \item{crossentropy}{The cross entropy error between \code{XProba} and
   \code{tcrossprod(Q, G)}.}
  \item{crossvalid.rmse}{If masked not \code{NULL}.
  The root square mean error between \code{XProba[masked]} and
  \code{tcrossprod(Q, G)[masked]}.}
   \item{crossvalid.crossentropy}{If masked not \code{NULL}.
   The cross entropy error between \code{XProba[masked]} and
   \code{tcrossprod(Q, G)[masked]}.}
}
}
\description{
\code{tess3Main} estimates spatial population structure using a graph based non
negative matrix factorization. After estimating the population structure is used to
compute a Fst statistic for each locus. See references for more details.
}
\examples{
library(tess3r)

# Arabidopsis thaliana data set
data(data.at)
genotype <- data.at$X
coordinates <- data.at$coord

# Run of tess3 main algorithm
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1)

# Run of tess3 main algorithm with cross validation errors computation.
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1,
                      mask = 0.05)


}
\references{
\url{https://hal.archives-ouvertes.fr/hal-01222555/} \url{http://biorxiv.org/content/early/2016/10/12/080291}
}
\seealso{
\code{\link{tess3}}
}

