% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{tess3Main}
\alias{tess3Main}
\title{Main function.}
\usage{
tess3Main(X, XProba = NULL, coord, K, ploidy, lambda = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, copy = TRUE,
  algo.copy = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{TODOC}

\item{coord}{TODOC}

\item{K}{TODOC}

\item{ploidy}{TODOC}

\item{lambda}{TODOC}

\item{W}{TODOC}

\item{method}{TODOC}

\item{max.iteration}{TODOC}

\item{tolerance}{TODOC}

\item{openMP.core.num}{TODOC}

\item{Q.init}{TODOC}

\item{mask}{TODOC}

\item{copy}{TODOC}

\item{algo.copy}{if TRUE data will be copy to speed the algorithm}

\item{verbose}{TODOC}

\item{XBin}{TODOC}
}
\value{
TODOC
}
\description{
Main function.
}

