% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3.R
\name{tess3Main}
\alias{tess3Main}
\title{Estimate ancestry coefficients and run genome scans for selection}
\usage{
tess3Main(X, XProba = NULL, coord, K, ploidy, lambda = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, copy = TRUE,
  algo.copy = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a numeric matrix which corresponds to the genotype matrix. This matrix
must be of size \eqn{n \times L} where \eqn{n} is the number of individuals and
\eqn{L} is the number of loci. Values of this matrix are integers corresponding
to the number of variant alleles observed at a locus. If \code{NULL}, \code{XProba}
is used.}

\item{XProba}{a numeric matrix which corresponds to genotype likelihoods (probabilities).
This matrix must be of size \eqn{n \times (ploidy + 1)L} where
\eqn{n} is the number of individuals and \eqn{L} is the number of loci. Entries of
this matrix are numeric values between 0 and 1 corresponding
to genotype probability. If \code{NULL}, this matrix is computed from the genotype matrix \code{X}.
See reference for details.}

\item{coord}{a numeric matrix of size \eqn{n \times 2} where \eqn{n} is the
number of individuals. It contains the geographic coordinates (Longitude, Latitude) of
all sampled individuals.}

\item{K}{an integer corresponding to
the number of ancestral populations.}

\item{ploidy}{an integer which corresponds to the ploidy of the number of copy of chromosomes.}

\item{lambda}{a nonnegative numeric which corresponds to the
spatial regularization parameter.}

\item{W}{a numeric matrix which corresponds to the graph weight matrix.
If \code{NULL}, W is computed as
\code{W[i,j] = exp(-(coord[i] - coord[j])^2 / sigma^2)},
where \code{coord[i]} is the set of geographic coordinates for individual i and
\code{sigma} equals 5 percent of the average geographic distance between individuals.}

\item{method}{\code{"projected.ls"} or \code{"qp"}. If \code{"projected.ls"},
an alternating projected least squares algorithm is used. If \code{"qp"},
an alternating quadratic programing algorithm is used. See references for details.}

\item{max.iteration}{the maximum number of
iterations in the optimization algorithm.}

\item{tolerance}{a numeric value which corresponds to the tolerance paramter in the
stopping criterion of the optimization algorithm.}

\item{openMP.core.num}{number of core used by the algorithm. It requires that openMP is
installed.}

\item{Q.init}{a numeric matrix which corresponds to the initial value of
\code{Q} for the algorithm.}

\item{mask}{if not \code{NULL}, this numeric value is the proportion of genotypic matrix entries
which are masked when computing the cross validation criterion.}

\item{copy}{if TRUE data will be copied once.}

\item{algo.copy}{if TRUE, data will be copied in order to speed the algorithm.}

\item{verbose}{If \code{TRUE} run information is printed.}
}
\value{
An object of class tess3Main which is a list with the following attributes:
\describe{
   \item{L}{the number of loci}
   \item{n}{the number of individuals}
   \item{ploidy}{the number of copies of chromosomes.}
   \item{K}{the number of ancestral populations.}
   \item{G}{the ancestral genotype frequency matrix.}
   \item{Q}{the ancestry coefficient matrix.}
   \item{Fst}{Fst statistic computed at each locus.}
   \item{Fscore}{Fscores computed from the Fst statistics.}
   \item{pvalue}{pvalues computed from the Fscores.}
   \item{log.pvalue}{The \eqn{log(pvalue)}.}
   \item{rmse}{root square mean error between \code{XProba} and
   \code{tcrossprod(Q, G)}.}
   \item{crossentropy}{cross-entropy error between \code{XProba} and
   \code{tcrossprod(Q, G)}.}
  \item{crossvalid.rmse}{if masked is not \code{NULL}, root square mean error
  between \code{XProba[masked]} and \code{tcrossprod(Q, G)[masked]}.}
   \item{crossvalid.crossentropy}{if masked not \code{NULL},
   the cross-entropy error between \code{XProba[masked]} and
   \code{tcrossprod(Q, G)[masked]}.}
}
}
\description{
\code{tess3Main} estimates spatial population structure using a graph based non
negative matrix factorization. After estimating the population structure is used to
compute a Fst statistic for each locus. See references for more details.
}
\examples{
library(tess3r)

# Arabidopsis thaliana data set
data(data.at)
genotype <- data.at$X
coordinates <- data.at$coord

# Run of tess3 main algorithm
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1)

# Run of tess3 main algorithm with cross validation errors computation.
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1,
                      mask = 0.05)


}
\references{
\url{https://hal.archives-ouvertes.fr/hal-01222555/} \url{http://biorxiv.org/content/early/2016/10/12/080291}
}
\seealso{
\code{\link{tess3}}
}
