% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3project.R
\name{tess3project}
\alias{tess3project}
\title{Title}
\usage{
tess3project(genotype, geographic.coordinate, K, ploidy, lambda, rep = 1,
  W = NULL, method = "MCPA", max.iteration = 50, tolerance = 1e-05,
  keep = "all")
}
\arguments{
\item{keep}{}
}
\description{
Title
}

